+++ demo/demo.js
@@ -127,8 +127,6 @@ app.controller('MainCtrl', function($scope, $rootScope, $timeout, $modal) {
     options: {
       ceil: 500,
       floor: 0,
-      step: 50,
-      showTicksValues: true,
       id: 'translate-slider',
       translate: function(value, id, which) {
         console.info(value, id, which);